+++ CHANGELOG.md
@@ -746,7 +746,10 @@
     paths to PS 7.6.6.
   * Updated script `Install-M365DSCAndDependencies` to work with a new PS
     version, new modules, etc.
-  * Added new ancillary script `Install-PowerShell` to be used by the Docker Windows images.
+  * Added new ancillary script `Install-PowerShell` to be used by the Docker
+    Windows images and the Linux dev image.
+  * Updated Docker files to work with latest code changes in preparation for
+    this breaking change.
 
 # 1.26.909.1